Machine Learning Project
Problem: To Mail or Not to Mail
Direct mailings to a company’s potential customers – “junk mail” to many – can be a very effective way for them
to market a product or a service. However, as we all know, much of this junk mail is really of no interest to the
people that receive it. Most of it ends up thrown away, not only wasting the money that the company spent on it,
but also filling up landfill waste sites or needing to be recycled.
If the company had a better understanding of who their potential customers were, they would know more
accurately who to send it to, so some of this waste and expense could be reduced.